py3: ported Fem to python3

+++ src/Mod/Fem/App/FemMeshPyImp.cpp
@@ -1,3 +1,4 @@
+
 /***************************************************************************
  *   Copyright (c) Jürgen Riegel          (juergen.riegel@web.de) 2009     *
  *                                                                         *
@@ -193,7 +194,7 @@ PyObject* FemMeshPy::addNode(PyObject *args)
             SMDS_MeshNode* node = meshDS->AddNode(x,y,z);
             if (!node)
                 throw std::runtime_error("Failed to add node");
-            return Py::new_reference_to(Py::Int(node->GetID()));
+            return Py::new_reference_to(Py::Long(node->GetID()));
         }
         catch (const std::exception& e) {
             PyErr_SetString(Base::BaseExceptionFreeCADError, e.what());
@@ -209,7 +210,7 @@ PyObject* FemMeshPy::addNode(PyObject *args)
             SMDS_MeshNode* node = meshDS->AddNodeWithID(x,y,z,i);
             if (!node)
                 throw std::runtime_error("Failed to add node");
-            return Py::new_reference_to(Py::Int(node->GetID()));
+            return Py::new_reference_to(Py::Long(node->GetID()));
         }
         catch (const std::exception& e) {
             PyErr_SetString(Base::BaseExceptionFreeCADError, e.what());
@@ -239,7 +240,7 @@ PyObject* FemMeshPy::addEdge(PyObject *args)
         SMDS_MeshEdge* edge = meshDS->AddEdge(node1, node2);
         if (!edge)
             throw std::runtime_error("Failed to add edge");
-        return Py::new_reference_to(Py::Int(edge->GetID()));
+        return Py::new_reference_to(Py::Long(edge->GetID()));
     }
     catch (const std::exception& e) {
         PyErr_SetString(Base::BaseExceptionFreeCADError, e.what());
@@ -265,7 +266,7 @@ PyObject* FemMeshPy::addFace(PyObject *args)
             SMDS_MeshFace* face = meshDS->AddFace(node1, node2, node3);
             if (!face)
                 throw std::runtime_error("Failed to add face");
-            return Py::new_reference_to(Py::Int(face->GetID()));
+            return Py::new_reference_to(Py::Long(face->GetID()));
         }
         catch (const std::exception& e) {
             PyErr_SetString(Base::BaseExceptionFreeCADError, e.what());
@@ -281,7 +282,7 @@ PyObject* FemMeshPy::addFace(PyObject *args)
         Py::List list(obj);
         std::vector<const SMDS_MeshNode*> Nodes;
         for (Py::List::iterator it = list.begin(); it != list.end(); ++it) {
-            Py::Int NoNr(*it);
+            Py::Long NoNr(*it);
             const SMDS_MeshNode* node = meshDS->FindNode(NoNr);
             if (!node)
                 throw std::runtime_error("Failed to get node of the given indices");
@@ -314,7 +315,7 @@ PyObject* FemMeshPy::addFace(PyObject *args)
                 throw std::runtime_error("Unknown node count, [3|4|6|8] are allowed"); //unknown face type
         }
 
-        return Py::new_reference_to(Py::Int(face->GetID()));
+        return Py::new_reference_to(Py::Long(face->GetID()));
 
     }
 
@@ -342,7 +343,7 @@ PyObject* FemMeshPy::addQuad(PyObject *args)
         SMDS_MeshFace* face = meshDS->AddFace(node1, node2, node3, node4);
         if (!face)
             throw std::runtime_error("Failed to add quad");
-        return Py::new_reference_to(Py::Int(face->GetID()));
+        return Py::new_reference_to(Py::Long(face->GetID()));
     }
     catch (const std::exception& e) {
         PyErr_SetString(Base::BaseExceptionFreeCADError, e.what());
@@ -368,7 +369,7 @@ PyObject* FemMeshPy::addVolume(PyObject *args)
             SMDS_MeshVolume* vol = meshDS->AddVolume(node1, node2, node3, node4);
             if (!vol)
                 throw std::runtime_error("Failed to add volume");
-            return Py::new_reference_to(Py::Int(vol->GetID()));
+            return Py::new_reference_to(Py::Long(vol->GetID()));
         }
         catch (const std::exception& e) {
             PyErr_SetString(Base::BaseExceptionFreeCADError, e.what());
@@ -384,7 +385,7 @@ PyObject* FemMeshPy::addVolume(PyObject *args)
         Py::List list(obj);
         std::vector<const SMDS_MeshNode*> Nodes;
         for (Py::List::iterator it = list.begin(); it != list.end(); ++it) {
-            Py::Int NoNr(*it);
+            Py::Long NoNr(*it);
             const SMDS_MeshNode* node = meshDS->FindNode(NoNr);
             if (!node)
                 throw std::runtime_error("Failed to get node of the given indices");
@@ -483,7 +484,7 @@ PyObject* FemMeshPy::addVolume(PyObject *args)
 
         }
 
-        return Py::new_reference_to(Py::Int(vol->GetID()));
+        return Py::new_reference_to(Py::Long(vol->GetID()));
 
     }
 
@@ -594,8 +595,8 @@ PyObject* FemMeshPy::getVolumesByFace(PyObject *args)
         std::list<std::pair<int, int> > resultSet = getFemMeshPtr()->getVolumesByFace(fc);
         for (std::list<std::pair<int, int> >::const_iterator it = resultSet.begin();it!=resultSet.end();++it) {
             Py::Tuple vol_face(2);
-            vol_face.setItem(0, Py::Int(it->first));
-            vol_face.setItem(1, Py::Int(it->second));
+            vol_face.setItem(0, Py::Long(it->first));
+            vol_face.setItem(1, Py::Long(it->second));
             ret.append(vol_face);
         }
 
@@ -627,8 +628,8 @@ PyObject* FemMeshPy::getccxVolumesByFace(PyObject *args)
         std::map<int, int> resultSet = getFemMeshPtr()->getccxVolumesByFace(fc);
         for (std::map<int, int>::const_iterator it = resultSet.begin();it!=resultSet.end();++it) {
             Py::Tuple vol_face(2);
-            vol_face.setItem(0, Py::Int(it->first));
-            vol_face.setItem(1, Py::Int(it->second));
+            vol_face.setItem(0, Py::Long(it->first));
+            vol_face.setItem(1, Py::Long(it->second));
             ret.append(vol_face);
         }
 
@@ -676,7 +677,7 @@ PyObject* FemMeshPy::getNodesBySolid(PyObject *args)
         Py::List ret;
         std::set<int> resultSet = getFemMeshPtr()->getNodesBySolid(fc);
         for (std::set<int>::const_iterator it = resultSet.begin();it!=resultSet.end();++it)
-            ret.append(Py::Int(*it));
+            ret.append(Py::Long(*it));
 
         return Py::new_reference_to(ret);
 
@@ -704,7 +705,7 @@ PyObject* FemMeshPy::getNodesByFace(PyObject *args)
         Py::List ret;
         std::set<int> resultSet = getFemMeshPtr()->getNodesByFace(fc);
         for (std::set<int>::const_iterator it = resultSet.begin();it!=resultSet.end();++it)
-            ret.append(Py::Int(*it));
+            ret.append(Py::Long(*it));
 
         return Py::new_reference_to(ret);
 
@@ -732,7 +733,7 @@ PyObject* FemMeshPy::getNodesByEdge(PyObject *args)
         Py::List ret;
         std::set<int> resultSet = getFemMeshPtr()->getNodesByEdge(fc);
         for (std::set<int>::const_iterator it = resultSet.begin();it!=resultSet.end();++it)
-            ret.append(Py::Int(*it));
+            ret.append(Py::Long(*it));
 
         return Py::new_reference_to(ret);
 
@@ -760,7 +761,7 @@ PyObject* FemMeshPy::getNodesByVertex(PyObject *args)
         Py::List ret;
         std::set<int> resultSet = getFemMeshPtr()->getNodesByVertex(fc);
         for (std::set<int>::const_iterator it = resultSet.begin();it!=resultSet.end();++it)
-            ret.append(Py::Int(*it));
+            ret.append(Py::Long(*it));
 
         return Py::new_reference_to(ret);
 
@@ -783,7 +784,7 @@ PyObject* FemMeshPy::getElementNodes(PyObject *args)
         Py::Tuple ret(resultSet.size());
         int index = 0;
         for (std::list<int>::const_iterator it = resultSet.begin();it!=resultSet.end();++it)
-            ret.setItem(index++, Py::Int(*it));
+            ret.setItem(index++, Py::Long(*it));
 
         return Py::new_reference_to(ret);
     }
@@ -814,15 +815,15 @@ Py::Dict FemMeshPy::getNodes(void) const
         vec = Mtrx * vec;
         int id = aNode->GetID();
 
-        dict[Py::Int(id)] = Py::asObject(new Base::VectorPy( vec ));
+        dict[Py::Long(id)] = Py::asObject(new Base::VectorPy( vec ));
     }
 
     return dict;
 }
 
-Py::Int FemMeshPy::getNodeCount(void) const
+Py::Long FemMeshPy::getNodeCount(void) const
 {
-    return Py::Int(getFemMeshPtr()->getSMesh()->NbNodes());
+    return Py::Long(getFemMeshPtr()->getSMesh()->NbNodes());
 }
 
 Py::Tuple FemMeshPy::getEdges(void) const
@@ -837,15 +838,15 @@ Py::Tuple FemMeshPy::getEdges(void) const
     Py::Tuple tuple(ids.size());
     int index = 0;
     for (std::set<int>::iterator it = ids.begin(); it != ids.end(); ++it) {
-        tuple.setItem(index++, Py::Int(*it));
+        tuple.setItem(index++, Py::Long(*it));
     }
 
     return tuple;
 }
 
-Py::Int FemMeshPy::getEdgeCount(void) const
+Py::Long FemMeshPy::getEdgeCount(void) const
 {
-    return Py::Int(getFemMeshPtr()->getSMesh()->NbEdges());
+    return Py::Long(getFemMeshPtr()->getSMesh()->NbEdges());
 }
 
 Py::Tuple FemMeshPy::getFaces(void) const
@@ -860,30 +861,30 @@ Py::Tuple FemMeshPy::getFaces(void) const
     Py::Tuple tuple(ids.size());
     int index = 0;
     for (std::set<int>::iterator it = ids.begin(); it != ids.end(); ++it) {
-        tuple.setItem(index++, Py::Int(*it));
+        tuple.setItem(index++, Py::Long(*it));
     }
 
     return tuple;
 }
 
-Py::Int FemMeshPy::getFaceCount(void) const
+Py::Long FemMeshPy::getFaceCount(void) const
 {
-    return Py::Int(getFemMeshPtr()->getSMesh()->NbFaces());
+    return Py::Long(getFemMeshPtr()->getSMesh()->NbFaces());
 }
 
-Py::Int FemMeshPy::getTriangleCount(void) const
+Py::Long FemMeshPy::getTriangleCount(void) const
 {
-    return Py::Int(getFemMeshPtr()->getSMesh()->NbTriangles());
+    return Py::Long(getFemMeshPtr()->getSMesh()->NbTriangles());
 }
 
-Py::Int FemMeshPy::getQuadrangleCount(void) const
+Py::Long FemMeshPy::getQuadrangleCount(void) const
 {
-    return Py::Int(getFemMeshPtr()->getSMesh()->NbQuadrangles());
+    return Py::Long(getFemMeshPtr()->getSMesh()->NbQuadrangles());
 }
 
-Py::Int FemMeshPy::getPolygonCount(void) const
+Py::Long FemMeshPy::getPolygonCount(void) const
 {
-    return Py::Int(getFemMeshPtr()->getSMesh()->NbPolygons());
+    return Py::Long(getFemMeshPtr()->getSMesh()->NbPolygons());
 }
 
 Py::Tuple FemMeshPy::getVolumes(void) const
@@ -898,50 +899,50 @@ Py::Tuple FemMeshPy::getVolumes(void) const
     Py::Tuple tuple(ids.size());
     int index = 0;
     for (std::set<int>::iterator it = ids.begin(); it != ids.end(); ++it) {
-        tuple.setItem(index++, Py::Int(*it));
+        tuple.setItem(index++, Py::Long(*it));
     }
 
     return tuple;
 }
 
-Py::Int FemMeshPy::getVolumeCount(void) const
+Py::Long FemMeshPy::getVolumeCount(void) const
 {
-    return Py::Int(getFemMeshPtr()->getSMesh()->NbVolumes());
+    return Py::Long(getFemMeshPtr()->getSMesh()->NbVolumes());
 }
 
-Py::Int FemMeshPy::getTetraCount(void) const
+Py::Long FemMeshPy::getTetraCount(void) const
 {
-    return Py::Int(getFemMeshPtr()->getSMesh()->NbTetras());
+    return Py::Long(getFemMeshPtr()->getSMesh()->NbTetras());
 }
 
-Py::Int FemMeshPy::getHexaCount(void) const
+Py::Long FemMeshPy::getHexaCount(void) const
 {
-    return Py::Int(getFemMeshPtr()->getSMesh()->NbHexas());
+    return Py::Long(getFemMeshPtr()->getSMesh()->NbHexas());
 }
 
-Py::Int FemMeshPy::getPyramidCount(void) const
+Py::Long FemMeshPy::getPyramidCount(void) const
 {
-    return Py::Int(getFemMeshPtr()->getSMesh()->NbPyramids());
+    return Py::Long(getFemMeshPtr()->getSMesh()->NbPyramids());
 }
 
-Py::Int FemMeshPy::getPrismCount(void) const
+Py::Long FemMeshPy::getPrismCount(void) const
 {
-    return Py::Int(getFemMeshPtr()->getSMesh()->NbPrisms());
+    return Py::Long(getFemMeshPtr()->getSMesh()->NbPrisms());
 }
 
-Py::Int FemMeshPy::getPolyhedronCount(void) const
+Py::Long FemMeshPy::getPolyhedronCount(void) const
 {
-    return Py::Int(getFemMeshPtr()->getSMesh()->NbPolyhedrons());
+    return Py::Long(getFemMeshPtr()->getSMesh()->NbPolyhedrons());
 }
 
-Py::Int FemMeshPy::getSubMeshCount(void) const
+Py::Long FemMeshPy::getSubMeshCount(void) const
 {
-    return Py::Int(getFemMeshPtr()->getSMesh()->NbSubMesh());
+    return Py::Long(getFemMeshPtr()->getSMesh()->NbSubMesh());
 }
 
-Py::Int FemMeshPy::getGroupCount(void) const
+Py::Long FemMeshPy::getGroupCount(void) const
 {
-    return Py::Int(getFemMeshPtr()->getSMesh()->NbGroup());
+    return Py::Long(getFemMeshPtr()->getSMesh()->NbGroup());
 }
 
 Py::Object FemMeshPy::getVolume(void) const
@@ -960,3 +961,4 @@ int FemMeshPy::setCustomAttributes(const char* /*attr*/, PyObject* /*obj*/)
 {
     return 0;
 }
+
